test: split workspace tests by responsibility
Keep shared fixtures in helpers_test.go and move create, remove, repository, rename, sync, status, doctor, and logging tests into focused files without changing test bodies.
