--- internal/workspace/workspace_test.go
@@ -1,2277 +1,0 @@
-package workspace
-
-import (
-	"bytes"
-	"errors"
-	"fmt"
-	"os"
-	"os/exec"
-	"path/filepath"
-	"strings"
-	"testing"
-	"time"
-
-	"github.com/nicksenap/grove/internal/gitops"
-	"github.com/nicksenap/grove/internal/logging"
-	"github.com/nicksenap/grove/internal/models"
-	"github.com/nicksenap/grove/internal/state"
-	"github.com/nicksenap/grove/internal/stats"
-)
-
-// testEnv sets up isolated Grove dirs + real git repos for testing.
-type testEnv struct {
-	t        *testing.T
-	dir      string
-	reposDir string
-	wsDir    string
-	groveDir string
-	cfg      *models.Config
-	repoMap  map[string]string
-	svc      *Service
-}
-
-func setupTestEnv(t *testing.T) *testEnv {
-	t.Helper()
-	dir := t.TempDir()
-
-	groveDir := filepath.Join(dir, ".grove")
-	wsDir := filepath.Join(groveDir, "workspaces")
-	reposDir := filepath.Join(dir, "repos")
-	os.MkdirAll(groveDir, 0o755)
-	os.MkdirAll(wsDir, 0o755)
-	os.MkdirAll(reposDir, 0o755)
-
-	store := state.NewStore(groveDir)
-	os.WriteFile(store.Path, []byte("[]"), 0o644)
-
-	cfg := &models.Config{
-		RepoDirs:     []string{reposDir},
-		WorkspaceDir: wsDir,
-		Presets:      map[string]models.Preset{},
-	}
-
-	svc := &Service{
-		State:        store,
-		Stats:        &stats.Tracker{StatsPath: filepath.Join(groveDir, "stats.json"), NowFn: time.Now},
-		RunCmd:       prodRunCmd,
-		RunCmdSilent: prodRunCmdSilent,
-	}
-
-	return &testEnv{
-		t:        t,
-		dir:      dir,
-		groveDir: groveDir,
-		reposDir: reposDir,
-		wsDir:    wsDir,
-		cfg:      cfg,
-		repoMap:  make(map[string]string),
-		svc:      svc,
-	}
-}
-
-func (e *testEnv) createWorkspace(name, branch string, repoNames []string) error {
-	e.t.Helper()
-	return e.svc.CreateWithOpts(name, CreateOpts{
-		Branch:  branch,
-		Repos:   repoNames,
-		RepoMap: e.repoMap,
-		Cfg:     e.cfg,
-	})
-}
-
-// createRepo creates a real git repo with an initial commit.
-func (e *testEnv) createRepo(name string) string {
-	e.t.Helper()
-	repoPath := filepath.Join(e.reposDir, name)
-	e.run(e.reposDir, "git", "init", "-q", repoPath)
-	e.run(repoPath, "git", "config", "user.email", "test@test.com")
-	e.run(repoPath, "git", "config", "user.name", "Test")
-	os.WriteFile(filepath.Join(repoPath, "README.md"), []byte("# "+name), 0o644)
-	e.run(repoPath, "git", "add", ".")
-	e.run(repoPath, "git", "commit", "-q", "-m", "initial")
-	e.repoMap[name] = repoPath
-	return repoPath
-}
-
-// createRepoWithRemote creates a repo cloned from a bare origin (for sync tests).
-func (e *testEnv) createRepoWithRemote(name string) string {
-	e.t.Helper()
-	bare := filepath.Join(e.reposDir, name+"-origin.git")
-	clone := filepath.Join(e.reposDir, name)
-
-	e.run(e.reposDir, "git", "init", "-q", "--bare", bare)
-	e.run(e.reposDir, "git", "clone", "-q", bare, clone)
-	e.run(clone, "git", "config", "user.email", "test@test.com")
-	e.run(clone, "git", "config", "user.name", "Test")
-	os.WriteFile(filepath.Join(clone, "README.md"), []byte("# "+name), 0o644)
-	e.run(clone, "git", "add", ".")
-	e.run(clone, "git", "commit", "-q", "-m", "initial")
-	e.run(clone, "git", "push", "-q", "origin", "HEAD")
-	e.repoMap[name] = clone
-	return clone
-}
-
-func (e *testEnv) run(dir string, name string, args ...string) string {
-	e.t.Helper()
-	cmd := exec.Command(name, args...)
-	cmd.Dir = dir
-	// Filter out GIT_DIR/GIT_WORK_TREE/GIT_INDEX_FILE to prevent leaking
-	// from parent process (e.g. when tests run inside a pre-commit hook).
-	for _, kv := range os.Environ() {
-		if strings.HasPrefix(kv, "GIT_DIR=") ||
-			strings.HasPrefix(kv, "GIT_WORK_TREE=") ||
-			strings.HasPrefix(kv, "GIT_INDEX_FILE=") {
-			continue
-		}
-		cmd.Env = append(cmd.Env, kv)
-	}
-	out, err := cmd.CombinedOutput()
-	if err != nil {
-		e.t.Fatalf("%s %s failed: %s\n%s", name, strings.Join(args, " "), err, out)
-	}
-	return strings.TrimSpace(string(out))
-}
-
-// ---------------------------------------------------------------------------
-// Create tests
-// ---------------------------------------------------------------------------
-
-func TestCreateSuccess(t *testing.T) {
-	env := setupTestEnv(t)
-	env.createRepo("api")
-	env.createRepo("web")
-
-	err := env.createWorkspace("test-ws", "feat/test", []string{"api", "web"})
-	if err != nil {
-		t.Fatalf("create: %v", err)
-	}
-
-	// Workspace saved to state
-	ws, _ := env.svc.State.GetWorkspace("test-ws")
-	if ws == nil {
-		t.Fatal("workspace not in state")
-	}
-	if len(ws.Repos) != 2 {
-		t.Errorf("expected 2 repos, got %d", len(ws.Repos))
-	}
-
-	// Worktree directories exist
-	for _, r := range ws.Repos {
-		if _, err := os.Stat(r.WorktreePath); os.IsNotExist(err) {
-			t.Errorf("worktree dir missing: %s", r.WorktreePath)
-		}
-	}
-
-	// Branch correct in worktrees
-	branch := env.run(filepath.Join(env.wsDir, "test-ws", "api"), "git", "branch", "--show-current")
-	if branch != "feat/test" {
-		t.Errorf("expected branch feat/test, got %s", branch)
-	}
-}
-
-func TestCreateWithPreparationUsesExactCommitOutsideLockAndSkipsSetup(t *testing.T) {
-	env := setupTestEnv(t)
-	repo := env.createRepo("api")
-	baseSHA := env.run(repo, "git", "rev-parse", "HEAD")
-	setupMarker := filepath.Join(env.dir, "legacy-setup-ran")
-	os.WriteFile(filepath.Join(repo, ".grove.toml"), []byte("setup = \"touch "+setupMarker+"\"\n"), 0o644)
-	os.WriteFile(filepath.Join(repo, "later.txt"), []byte("later"), 0o644)
-	env.run(repo, "git", "add", ".")
-	env.run(repo, "git", "commit", "-q", "-m", "later")
-
-	var preparedPath string
-	err := env.svc.CreateWithPreparation("prepared", PreparationOpts{
-		CreateOpts:  CreateOpts{Branch: "feat/prepared", Repos: []string{"api"}, RepoMap: env.repoMap, Cfg: env.cfg},
-		BaseCommits: map[string]string{"api": baseSHA},
-	}, func(ws models.Workspace) error {
-		if err := env.svc.State.WithLock(func() error { return nil }); err != nil {
-			return fmt.Errorf("acquiring state lock from preparation: %w", err)
-		}
-		preparedPath = ws.Repos[0].WorktreePath
-		if got := env.run(preparedPath, "git", "rev-parse", "HEAD"); got != baseSHA {
-			return fmt.Errorf("worktree HEAD = %s, want %s", got, baseSHA)
-		}
-		return nil
-	})
-	if err != nil {
-		t.Fatal(err)
-	}
-	if preparedPath == "" {
-		t.Fatal("preparation callback did not run")
-	}
-	if _, err := os.Stat(setupMarker); !os.IsNotExist(err) {
-		t.Fatalf("legacy setup hook ran during Recipe creation: %v", err)
-	}
-}
-
-func TestCreateWithPreparationFailurePreservesPreexistingBranch(t *testing.T) {
-	env := setupTestEnv(t)
-	repo := env.createRepo("api")
-	baseSHA := env.run(repo, "git", "rev-parse", "HEAD")
-	env.run(repo, "git", "branch", "feat/existing", baseSHA)
-	boom := errors.New("prepare failed")
-
-	err := env.svc.CreateWithPreparation("prepared", PreparationOpts{
-		CreateOpts:  CreateOpts{Branch: "feat/existing", Repos: []string{"api"}, RepoMap: env.repoMap, Cfg: env.cfg},
-		BaseCommits: map[string]string{"api": baseSHA},
-	}, func(models.Workspace) error { return boom })
-	var preparationErr *PreparationError
-	if !errors.As(err, &preparationErr) || !errors.Is(err, boom) {
-		t.Fatalf("error = %v, want PreparationError wrapping callback failure", err)
-	}
-	if !gitops.BranchExists(repo, "feat/existing") {
-		t.Fatal("pre-existing branch was deleted")
-	}
-	if ws, _ := env.svc.State.GetWorkspace("prepared"); ws != nil {
-		t.Fatalf("failed preparation remained in state: %+v", ws)
-	}
-	if _, err := os.Stat(filepath.Join(env.wsDir, "prepared")); !os.IsNotExist(err) {
-		t.Fatalf("workspace directory remained after rollback: %v", err)
-	}
-}
-
-func TestDeletePreparedWorkspacePreservesPreexistingBranch(t *testing.T) {
-	env := setupTestEnv(t)
-	repo := env.createRepo("api")
-	baseSHA := env.run(repo, "git", "rev-parse", "HEAD")
-	env.run(repo, "git", "branch", "feat/existing", baseSHA)
-
-	if err := env.svc.CreateWithPreparation("prepared", PreparationOpts{
-		CreateOpts:  CreateOpts{Branch: "feat/existing", Repos: []string{"api"}, RepoMap: env.repoMap, Cfg: env.cfg},
-		BaseCommits: map[string]string{"api": baseSHA},
-	}, func(models.Workspace) error { return nil }); err != nil {
-		t.Fatal(err)
-	}
-	if err := env.svc.DeleteWithOptions("prepared", RemoveOptions{Force: true}); err != nil {
-		t.Fatal(err)
-	}
-	if !gitops.BranchExists(repo, "feat/existing") {
-		t.Fatal("normal workspace deletion removed pre-existing Recipe branch")
-	}
-}
-
-func TestCreateWithPreparationFailureRemovesOwnedBranch(t *testing.T) {
-	env := setupTestEnv(t)
-	repo := env.createRepo("api")
-	baseSHA := env.run(repo, "git", "rev-parse", "HEAD")
-
-	_ = env.svc.CreateWithPreparation("prepared", PreparationOpts{
-		CreateOpts:  CreateOpts{Branch: "feat/owned", Repos: []string{"api"}, RepoMap: env.repoMap, Cfg: env.cfg},
-		BaseCommits: map[string]string{"api": baseSHA},
-	}, func(models.Workspace) error { return errors.New("prepare failed") })
-	if gitops.BranchExists(repo, "feat/owned") {
-		t.Fatal("branch created by failed preparation was preserved")
-	}
-}
-
-func TestCreateWithPreparationRejectsSuccessfulIdentityChange(t *testing.T) {
-	env := setupTestEnv(t)
-	repo := env.createRepo("api")
-	baseSHA := env.run(repo, "git", "rev-parse", "HEAD")
-
-	err := env.svc.CreateWithPreparation("prepared", PreparationOpts{
-		CreateOpts:  CreateOpts{Branch: "feat/prepared", Repos: []string{"api"}, RepoMap: env.repoMap, Cfg: env.cfg},
-		BaseCommits: map[string]string{"api": baseSHA},
-	}, func(ws models.Workspace) error {
-		env.run(ws.Repos[0].WorktreePath, "git", "checkout", "--detach")
-		return nil
-	})
-	var preparationErr *PreparationError
-	if !errors.As(err, &preparationErr) || preparationErr.CleanupErr == nil {
-		t.Fatalf("error = %v, want identity verification failure", err)
-	}
-	if ws, _ := env.svc.State.GetWorkspace("prepared"); ws == nil {
-		t.Fatal("changed worktree state should remain recoverable")
-	}
-}
-
-func TestCreateWithPreparationRefusesRollbackAfterWorktreeIdentityChanges(t *testing.T) {
-	env := setupTestEnv(t)
-	repo := env.createRepo("api")
-	baseSHA := env.run(repo, "git", "rev-parse", "HEAD")
-
-	err := env.svc.CreateWithPreparation("prepared", PreparationOpts{
-		CreateOpts:  CreateOpts{Branch: "feat/prepared", Repos: []string{"api"}, RepoMap: env.repoMap, Cfg: env.cfg},
-		BaseCommits: map[string]string{"api": baseSHA},
-	}, func(ws models.Workspace) error {
-		env.run(ws.Repos[0].WorktreePath, "git", "checkout", "--detach")
-		return errors.New("prepare failed")
-	})
-	var preparationErr *PreparationError
-	if !errors.As(err, &preparationErr) || preparationErr.CleanupErr == nil {
-		t.Fatalf("error = %v, want identity cleanup refusal", err)
-	}
-	if ws, _ := env.svc.State.GetWorkspace("prepared"); ws == nil {
-		t.Fatal("workspace state was removed after worktree identity changed")
-	}
-	if _, err := os.Stat(filepath.Join(env.wsDir, "prepared", "api")); err != nil {
-		t.Fatalf("changed worktree was removed: %v", err)
-	}
-}
-
-func TestCreateWithPreparationRefusesRollbackAfterStateChanges(t *testing.T) {
-	env := setupTestEnv(t)
-	repo := env.createRepo("api")
-	baseSHA := env.run(repo, "git", "rev-parse", "HEAD")
-
-	err := env.svc.CreateWithPreparation("prepared", PreparationOpts{
-		CreateOpts:  CreateOpts{Branch: "feat/prepared", Repos: []string{"api"}, RepoMap: env.repoMap, Cfg: env.cfg},
-		BaseCommits: map[string]string{"api": baseSHA},
-	}, func(ws models.Workspace) error {
-		ws.Branch = "concurrently-changed"
-		if err := env.svc.State.UpdateWorkspace(ws); err != nil {
-			return err
-		}
-		return errors.New("prepare failed")
-	})
-	var preparationErr *PreparationError
-	if !errors.As(err, &preparationErr) || preparationErr.CleanupErr == nil {
-		t.Fatalf("error = %v, want cleanup refusal", err)
-	}
-	if ws, _ := env.svc.State.GetWorkspace("prepared"); ws == nil || ws.Branch != "concurrently-changed" {
-		t.Fatalf("changed workspace state was removed: %+v", ws)
-	}
-	if _, err := os.Stat(filepath.Join(env.wsDir, "prepared", "api")); err != nil {
-		t.Fatalf("worktree was removed despite changed state: %v", err)
-	}
-}
-
-func TestCreateDuplicateNameFails(t *testing.T) {
-	env := setupTestEnv(t)
-	env.createRepo("api")
-
-	env.createWorkspace("dupe", "feat/a", []string{"api"})
-
-	err := env.createWorkspace("dupe", "feat/b", []string{"api"})
-	if err == nil {
-		t.Error("expected error for duplicate workspace name")
-	}
-}
-
-func TestCreateDuplicateBranchFails(t *testing.T) {
-	env := setupTestEnv(t)
-	env.createRepo("api")
-
-	env.createWorkspace("ws1", "feat/shared", []string{"api"})
-
-	// Second workspace with same branch on same repo should fail
-	err := env.createWorkspace("ws2", "feat/shared", []string{"api"})
-	if err == nil {
-		t.Error("expected error for duplicate branch on same repo")
-	}
-}
-
-func TestCreateRollbackOnFailure(t *testing.T) {
-	env := setupTestEnv(t)
-	env.createRepo("api")
-
-	// Try to create with a nonexistent repo — should rollback
-	err := env.createWorkspace("rollback-ws", "feat/test", []string{"api", "nonexistent"})
-	if err == nil {
-		t.Error("expected error")
-	}
-
-	// Workspace dir should be cleaned up
-	wsPath := filepath.Join(env.wsDir, "rollback-ws")
-	if _, err := os.Stat(wsPath); !os.IsNotExist(err) {
-		t.Error("workspace dir should be removed on rollback")
-	}
-
-	// State should not contain the workspace
-	ws, _ := env.svc.State.GetWorkspace("rollback-ws")
-	if ws != nil {
-		t.Error("workspace should not be in state after rollback")
-	}
-}
-
-func TestCreateFailurePreservesPreexistingWorkspaceRoot(t *testing.T) {
-	env := setupTestEnv(t)
-	path := filepath.Join(env.wsDir, "preexisting-root")
-	if err := os.Mkdir(path, 0o755); err != nil {
-		t.Fatalf("create pre-existing root: %v", err)
-	}
-
-	err := env.createWorkspace("preexisting-root", "feat/preexisting-root", []string{"missing"})
-	if err == nil {
-		t.Fatal("expected missing repo error")
-	}
-	if _, err := os.Stat(path); err != nil {
-		t.Fatalf("pre-existing workspace root should remain: %v", err)
-	}
-}
-
-func TestCreateRollbackRemovesBranchesCreatedByInvocation(t *testing.T) {
-	env := setupTestEnv(t)
-	api := env.createRepo("api")
-	web := env.createRepo("web")
-
-	conflictPath := filepath.Join(env.dir, "web-conflict")
-	env.run(web, "git", "worktree", "add", "-q", "-b", "feat/create-rollback", conflictPath)
-
-	err := env.createWorkspace("create-rollback", "feat/create-rollback", []string{"api", "web"})
-	if err == nil || !strings.Contains(err.Error(), "web") {
-		t.Fatalf("expected web conflict, got %v", err)
-	}
-	if gitops.BranchExists(api, "feat/create-rollback") {
-		t.Error("rollback should remove the api branch created by this invocation")
-	}
-	if _, err := os.Stat(filepath.Join(env.wsDir, "create-rollback")); !os.IsNotExist(err) {
-		t.Error("empty workspace root should be removed after rollback")
-	}
-}
-
-func TestConcurrentCreatesPreserveBothStateEntries(t *testing.T) {
-	env := setupTestEnv(t)
-	env.createRepo("api")
-	env.createRepo("web")
-
-	svcA := *env.svc
-	svcA.State = state.NewStore(env.groveDir)
-	svcA.Stats = &stats.Tracker{StatsPath: filepath.Join(env.groveDir, "stats-a.json"), NowFn: time.Now}
-	svcB := *env.svc
-	svcB.State = state.NewStore(env.groveDir)
-	svcB.Stats = &stats.Tracker{StatsPath: filepath.Join(env.groveDir, "stats-b.json"), NowFn: time.Now}
-
-	start := make(chan struct{})
-	errs := make(chan error, 2)
-	go func() {
-		<-start
-		errs <- svcA.CreateWithOpts("ws-a", CreateOpts{
-			Branch: "feat/a", Repos: []string{"api"}, RepoMap: env.repoMap, Cfg: env.cfg,
-		})
-	}()
-	go func() {
-		<-start
-		errs <- svcB.CreateWithOpts("ws-b", CreateOpts{
-			Branch: "feat/b", Repos: []string{"web"}, RepoMap: env.repoMap, Cfg: env.cfg,
-		})
-	}()
-	close(start)
-
-	for range 2 {
-		if err := <-errs; err != nil {
-			t.Fatalf("concurrent create: %v", err)
-		}
-	}
-
-	workspaces, err := env.svc.State.Load()
-	if err != nil {
-		t.Fatalf("load state: %v", err)
-	}
-	if len(workspaces) != 2 {
-		t.Fatalf("expected both workspaces in state, got %d", len(workspaces))
-	}
-}
-
-func TestCreateAutoCreatesBranch(t *testing.T) {
-	env := setupTestEnv(t)
-	env.createRepo("api")
-
-	// Branch doesn't exist yet — should be auto-created
-	err := env.createWorkspace("auto-branch", "feat/new-branch", []string{"api"})
-	if err != nil {
-		t.Fatalf("create: %v", err)
-	}
-
-	// Verify the branch exists in the worktree
-	branch := env.run(filepath.Join(env.wsDir, "auto-branch", "api"), "git", "branch", "--show-current")
-	if branch != "feat/new-branch" {
-		t.Errorf("expected feat/new-branch, got %s", branch)
-	}
-}
-
-// pushRemoteBranch creates a branch in repo, commits a marker file, pushes it to
-// origin, then deletes the local branch — leaving only origin/<branch>. This
-// simulates a PR head branch that exists on the remote but not locally.
-func (e *testEnv) pushRemoteBranch(repo, branch, marker string) {
-	e.t.Helper()
-	e.run(repo, "git", "checkout", "-q", "-b", branch)
-	os.WriteFile(filepath.Join(repo, marker), []byte(marker), 0o644)
-	e.run(repo, "git", "add", ".")
-	e.run(repo, "git", "commit", "-q", "-m", "remote work on "+branch)
-	e.run(repo, "git", "push", "-q", "origin", branch)
-	e.run(repo, "git", "checkout", "-q", "-")
-	e.run(repo, "git", "branch", "-q", "-D", branch)
-	// Forget the just-pushed ref locally so creation must rely on fetch.
-	e.run(repo, "git", "update-ref", "-d", "refs/remotes/origin/"+branch)
-}
-
-func TestCreateTrackModeChecksOutExistingBranch(t *testing.T) {
-	env := setupTestEnv(t)
-	env.createRepoWithRemote("api")
-	env.pushRemoteBranch(env.repoMap["api"], "feat/pr-head", "pr-marker.txt")
-
-	err := env.svc.CreateWithOpts("pr-ws", CreateOpts{
-		Branch:          "feat/pr-head",
-		Repos:           []string{"api"},
-		RepoMap:         env.repoMap,
-		Cfg:             env.cfg,
-		BranchMode:      BranchModeTrack,
-		TrackBranchRepo: "api",
-	})
-	if err != nil {
-		t.Fatalf("create track: %v", err)
-	}
-
-	wt := filepath.Join(env.wsDir, "pr-ws", "api")
-	// The PR head's marker file must be present (we checked out the existing branch).
-	if _, err := os.Stat(filepath.Join(wt, "pr-marker.txt")); os.IsNotExist(err) {
-		t.Error("expected PR head content (pr-marker.txt) in tracking worktree")
-	}
-	// The worktree branch must track origin/feat/pr-head.
-	upstream := env.run(wt, "git", "rev-parse", "--abbrev-ref", "--symbolic-full-name", "@{upstream}")
-	if upstream != "origin/feat/pr-head" {
-		t.Errorf("expected upstream origin/feat/pr-head, got %q", upstream)
-	}
-}
-
-func TestCreateTrackModeFallsBackWhenRemoteMissing(t *testing.T) {
-	env := setupTestEnv(t)
-	env.createRepoWithRemote("api")
-
-	// Track mode requested but no such remote branch exists → fall back to
-	// creating a new branch from base (no error).
-	err := env.svc.CreateWithOpts("fallback-ws", CreateOpts{
-		Branch:          "feat/ghost-pr",
-		Repos:           []string{"api"},
-		RepoMap:         env.repoMap,
-		Cfg:             env.cfg,
-		BranchMode:      BranchModeTrack,
-		TrackBranchRepo: "api",
-	})
-	if err != nil {
-		t.Fatalf("expected graceful fallback, got error: %v", err)
-	}
-
-	wt := filepath.Join(env.wsDir, "fallback-ws", "api")
-	branch := env.run(wt, "git", "branch", "--show-current")
-	if branch != "feat/ghost-pr" {
-		t.Errorf("expected new branch feat/ghost-pr, got %q", branch)
-	}
-}
-
-func TestCreateTrackModeOnlyAppliesToDesignatedRepo(t *testing.T) {
-	env := setupTestEnv(t)
-	env.createRepoWithRemote("api")
-	env.createRepoWithRemote("web")
-	env.pushRemoteBranch(env.repoMap["api"], "feat/shared", "api-pr.txt")
-
-	// Both repos share the branch name, but only "api" is the track repo;
-	// "web" should create a fresh branch from base (no api-pr.txt leakage).
-	err := env.svc.CreateWithOpts("mixed-ws", CreateOpts{
-		Branch:          "feat/shared",
-		Repos:           []string{"api", "web"},
-		RepoMap:         env.repoMap,
-		Cfg:             env.cfg,
-		BranchMode:      BranchModeTrack,
-		TrackBranchRepo: "api",
-	})
-	if err != nil {
-		t.Fatalf("create mixed: %v", err)
-	}
-
-	apiWT := filepath.Join(env.wsDir, "mixed-ws", "api")
-	if _, err := os.Stat(filepath.Join(apiWT, "api-pr.txt")); os.IsNotExist(err) {
-		t.Error("api worktree should have tracked the PR head (api-pr.txt missing)")
-	}
-	webWT := filepath.Join(env.wsDir, "mixed-ws", "web")
-	if _, err := os.Stat(filepath.Join(webWT, "api-pr.txt")); err == nil {
-		t.Error("web worktree should NOT contain api's PR content — create mode expected")
-	}
-}
-
-func TestCreateWithOptsPersistsSource(t *testing.T) {
-	env := setupTestEnv(t)
-	env.createRepo("api")
-
-	src := &models.WorkspaceSource{
-		Provider: "github",
-		URL:      "https://github.com/acme/example-app/pull/1172",
-		Ref:      "1172",
-		Title:    "Example pull request",
-	}
-	err := env.svc.CreateWithOpts("src-ws", CreateOpts{
-		Branch:  "feat/src",
-		Repos:   []string{"api"},
-		RepoMap: env.repoMap,
-		Cfg:     env.cfg,
-		Source:  src,
-	})
-	if err != nil {
-		t.Fatalf("create: %v", err)
-	}
-
-	ws, _ := env.svc.State.GetWorkspace("src-ws")
-	if ws == nil || ws.Source == nil {
-		t.Fatal("expected persisted Source on workspace")
-	}
-	if *ws.Source != *src {
-		t.Errorf("source mismatch: got %+v, want %+v", *ws.Source, *src)
-	}
-}
-
-func TestCreateDoesNotWriteMCPConfig(t *testing.T) {
-	env := setupTestEnv(t)
-	env.createRepo("api")
-
-	if err := env.createWorkspace("plain-ws", "feat/plain", []string{"api"}); err != nil {
-		t.Fatalf("create: %v", err)
-	}
-
-	mcpPath := filepath.Join(env.wsDir, "plain-ws", ".mcp.json")
-	if _, err := os.Stat(mcpPath); !os.IsNotExist(err) {
-		t.Fatalf("Grove should not create .mcp.json, stat error: %v", err)
-	}
-}
-
-func TestCreateCdFile(t *testing.T) {
-	env := setupTestEnv(t)
-	env.createRepo("api")
-
-	cdFile := filepath.Join(env.dir, "cd-target")
-	os.Setenv("GROVE_CD_FILE", cdFile)
-	defer os.Unsetenv("GROVE_CD_FILE")
-
-	env.createWorkspace("cd-ws", "feat/cd", []string{"api"})
-
-	data, err := os.ReadFile(cdFile)
-	if err != nil {
-		t.Fatalf("reading cd file: %v", err)
-	}
-	expected := filepath.Join(env.wsDir, "cd-ws")
-	if string(data) != expected {
-		t.Errorf("cd file: got %q, want %q", string(data), expected)
-	}
-}
-
-// ---------------------------------------------------------------------------
-// Setup hook tests
-// ---------------------------------------------------------------------------
-
-func TestSetupHookRuns(t *testing.T) {
-	env := setupTestEnv(t)
-	repo := env.createRepo("hooked")
-
-	// Write .grove.toml with setup hook
-	toml := `setup = "touch .setup-ran"`
-	os.WriteFile(filepath.Join(repo, ".grove.toml"), []byte(toml), 0o644)
-	env.run(repo, "git", "add", ".")
-	env.run(repo, "git", "commit", "-q", "-m", "add grove config")
-
-	env.createWorkspace("hook-ws", "feat/hook", []string{"hooked"})
-
-	// Check marker file in worktree (not source repo)
-	marker := filepath.Join(env.wsDir, "hook-ws", "hooked", ".setup-ran")
-	if _, err := os.Stat(marker); os.IsNotExist(err) {
-		t.Error("setup hook did not run in worktree")
-	}
-}
-
-func TestSetupHookMultipleCommands(t *testing.T) {
-	env := setupTestEnv(t)
-	repo := env.createRepo("multi")
-
-	toml := `setup = ["touch .step1", "touch .step2"]`
-	os.WriteFile(filepath.Join(repo, ".grove.toml"), []byte(toml), 0o644)
-	env.run(repo, "git", "add", ".")
-	env.run(repo, "git", "commit", "-q", "-m", "add grove config")
-
-	env.createWorkspace("multi-ws", "feat/multi", []string{"multi"})
-
-	wt := filepath.Join(env.wsDir, "multi-ws", "multi")
-	for _, f := range []string{".step1", ".step2"} {
-		if _, err := os.Stat(filepath.Join(wt, f)); os.IsNotExist(err) {
-			t.Errorf("setup hook step %s did not run", f)
-		}
-	}
-}
-
-func TestNoSetupHookNoCrash(t *testing.T) {
-	env := setupTestEnv(t)
-	env.createRepo("plain")
-
-	err := env.createWorkspace("plain-ws", "feat/plain", []string{"plain"})
-	if err != nil {
-		t.Fatalf("should not fail without setup hook: %v", err)
-	}
-}
-
-// ---------------------------------------------------------------------------
-// Delete tests
-// ---------------------------------------------------------------------------
-
-func TestDeleteSuccess(t *testing.T) {
-	env := setupTestEnv(t)
-	env.createRepo("api")
-	env.createWorkspace("del-ws", "feat/del", []string{"api"})
-
-	err := env.svc.Delete("del-ws")
-	if err != nil {
-		t.Fatalf("delete: %v", err)
-	}
-
-	// State cleared
-	ws, _ := env.svc.State.GetWorkspace("del-ws")
-	if ws != nil {
-		t.Error("workspace should be removed from state")
-	}
-
-	// Directory cleaned up
-	if _, err := os.Stat(filepath.Join(env.wsDir, "del-ws")); !os.IsNotExist(err) {
-		t.Error("workspace dir should be removed")
-	}
-}
-
-func TestDeleteNotFound(t *testing.T) {
-	env := setupTestEnv(t)
-	_ = env // setup env for state path
-
-	err := env.svc.Delete("nonexistent")
-	if err == nil {
-		t.Error("expected error for nonexistent workspace")
-	}
-}
-
-func TestDeleteCleansBranch(t *testing.T) {
-	env := setupTestEnv(t)
-	env.createRepo("api")
-	env.createWorkspace("branch-ws", "feat/to-clean", []string{"api"})
-
-	env.svc.Delete("branch-ws")
-
-	// Branch should be cleaned up from source repo
-	out := env.run(env.repoMap["api"], "git", "branch", "--list", "feat/to-clean")
-	if strings.TrimSpace(out) != "" {
-		t.Error("branch should be deleted from source repo")
-	}
-}
-
-func TestDeleteRejectsDirtyWorktreeByDefault(t *testing.T) {
-	env := setupTestEnv(t)
-	env.createRepo("api")
-	if err := env.createWorkspace("dirty-delete", "feat/dirty-delete", []string{"api"}); err != nil {
-		t.Fatalf("create: %v", err)
-	}
-
-	ws, _ := env.svc.State.GetWorkspace("dirty-delete")
-	dirtyFile := filepath.Join(ws.Repos[0].WorktreePath, "dirty.txt")
-	if err := os.WriteFile(dirtyFile, []byte("keep me"), 0o644); err != nil {
-		t.Fatalf("write dirty file: %v", err)
-	}
-
-	err := env.svc.DeleteWithOptions("dirty-delete", RemoveOptions{})
-	if err == nil || !strings.Contains(err.Error(), "api") || !strings.Contains(err.Error(), "dirty") {
-		t.Fatalf("expected repo-named dirty error, got %v", err)
-	}
-	if _, err := os.Stat(dirtyFile); err != nil {
-		t.Fatalf("dirty file should remain: %v", err)
-	}
-	if saved, _ := env.svc.State.GetWorkspace("dirty-delete"); saved == nil {
-		t.Fatal("workspace state should remain")
-	}
-}
-
-func TestDeleteRemovesWorkspaceMetadata(t *testing.T) {
-	env := setupTestEnv(t)
-	env.createRepo("api")
-	if err := env.createWorkspace("metadata-delete", "feat/metadata-delete", []string{"api"}); err != nil {
-		t.Fatalf("create: %v", err)
-	}
-
-	ws, _ := env.svc.State.GetWorkspace("metadata-delete")
-	if err := os.MkdirAll(filepath.Join(ws.Path, ".pi"), 0o755); err != nil {
-		t.Fatalf("create workspace metadata directory: %v", err)
-	}
-	if err := os.WriteFile(filepath.Join(ws.Path, ".pi", "hindsight.json"), []byte("{}"), 0o644); err != nil {
-		t.Fatalf("write workspace metadata: %v", err)
-	}
-
-	if err := env.svc.Delete("metadata-delete"); err != nil {
-		t.Fatalf("delete: %v", err)
-	}
-	if _, err := os.Stat(ws.Path); !os.IsNotExist(err) {
-		t.Fatalf("workspace metadata should not block root removal, got %v", err)
-	}
-}
-
-func TestDeleteForceRemovesDirtyWorktree(t *testing.T) {
-	env := setupTestEnv(t)
-	env.createRepo("api")
-	if err := env.createWorkspace("force-delete", "feat/force-delete", []string{"api"}); err != nil {
-		t.Fatalf("create: %v", err)
-	}
-
-	ws, _ := env.svc.State.GetWorkspace("force-delete")
-	if err := os.WriteFile(filepath.Join(ws.Repos[0].WorktreePath, "dirty.txt"), []byte("remove me"), 0o644); err != nil {
-		t.Fatalf("write dirty file: %v", err)
-	}
-	if err := os.MkdirAll(filepath.Join(ws.Path, ".pi"), 0o755); err != nil {
-		t.Fatalf("create workspace metadata directory: %v", err)
-	}
-	if err := os.WriteFile(filepath.Join(ws.Path, ".pi", "hindsight.json"), []byte("{}"), 0o644); err != nil {
-		t.Fatalf("write workspace metadata: %v", err)
-	}
-
-	if err := env.svc.DeleteWithOptions("force-delete", RemoveOptions{Force: true}); err != nil {
-		t.Fatalf("force delete: %v", err)
-	}
-	if saved, _ := env.svc.State.GetWorkspace("force-delete"); saved != nil {
-		t.Fatal("workspace state should be removed")
-	}
-	if _, err := os.Stat(ws.Path); !os.IsNotExist(err) {
-		t.Fatalf("force delete should remove the entire workspace root, got %v", err)
-	}
-}
-
-func TestDeleteRemovalFailurePreservesPathAndState(t *testing.T) {
-	env := setupTestEnv(t)
-	env.createRepo("api")
-	if err := env.createWorkspace("failed-delete", "feat/failed-delete", []string{"api"}); err != nil {
-		t.Fatalf("create: %v", err)
-	}
-
-	ws, _ := env.svc.State.GetWorkspace("failed-delete")
-	env.svc.RemoveWorktree = func(repo, path string, force bool) error {
-		return fmt.Errorf("simulated git failure")
-	}
-
-	err := env.svc.Delete("failed-delete")
-	if err == nil || !strings.Contains(err.Error(), "api") {
-		t.Fatalf("expected repo-named removal error, got %v", err)
-	}
-	if _, err := os.Stat(ws.Repos[0].WorktreePath); err != nil {
-		t.Fatalf("worktree path should remain: %v", err)
-	}
-	if saved, _ := env.svc.State.GetWorkspace("failed-delete"); saved == nil || len(saved.Repos) != 1 {
-		t.Fatal("failed repo should remain in state")
-	}
-}
-
-func TestDeletePreflightsEveryRepoBeforeRemovingAny(t *testing.T) {
-	env := setupTestEnv(t)
-	env.createRepo("api")
-	env.createRepo("web")
-	if err := env.createWorkspace("preflight-delete", "feat/preflight-delete", []string{"api", "web"}); err != nil {
-		t.Fatalf("create: %v", err)
-	}
-
-	ws, _ := env.svc.State.GetWorkspace("preflight-delete")
-	apiPath := ws.Repos[0].WorktreePath
-	ws.Repos[1].WorktreePath = filepath.Join(ws.Path, "unexpected-web")
-	if err := env.svc.State.UpdateWorkspace(*ws); err != nil {
-		t.Fatalf("corrupt state fixture: %v", err)
-	}
-
-	err := env.svc.Delete("preflight-delete")
-	if err == nil || !strings.Contains(err.Error(), "web") {
-		t.Fatalf("expected web preflight error, got %v", err)
-	}
-	if _, err := os.Stat(apiPath); err != nil {
-		t.Fatalf("api should not be removed before all preflights pass: %v", err)
-	}
-}
-
-// ---------------------------------------------------------------------------
-// Rename tests
-// ---------------------------------------------------------------------------
-
-func TestRenameSuccess(t *testing.T) {
-	env := setupTestEnv(t)
-	env.createRepo("api")
-	env.createWorkspace("old-name", "feat/rename", []string{"api"})
-
-	err := env.svc.Rename("old-name", "new-name")
-	if err != nil {
-		t.Fatalf("rename: %v", err)
-	}
-
-	// Old name gone
-	ws, _ := env.svc.State.GetWorkspace("old-name")
-	if ws != nil {
-		t.Error("old name should not exist")
-	}
-
-	// New name exists with updated paths
-	ws, _ = env.svc.State.GetWorkspace("new-name")
-	if ws == nil {
-		t.Fatal("new workspace not found")
-	}
-	if !strings.Contains(ws.Path, "new-name") {
-		t.Errorf("path should contain new-name: %s", ws.Path)
-	}
-	for _, r := range ws.Repos {
-		if !strings.Contains(r.WorktreePath, "new-name") {
-			t.Errorf("worktree path should contain new-name: %s", r.WorktreePath)
-		}
-	}
-
-	// Directory renamed
-	if _, err := os.Stat(filepath.Join(env.wsDir, "new-name")); os.IsNotExist(err) {
-		t.Error("new directory should exist")
-	}
-	if _, err := os.Stat(filepath.Join(env.wsDir, "old-name")); !os.IsNotExist(err) {
-		t.Error("old directory should not exist")
-	}
-}
-
-func TestRenameNotFound(t *testing.T) {
-	env := setupTestEnv(t)
-	_ = env
-
-	err := env.svc.Rename("nonexistent", "new")
-	if err == nil {
-		t.Error("expected error")
-	}
-}
-
-func TestRenameNameTaken(t *testing.T) {
-	env := setupTestEnv(t)
-	env.createRepo("api")
-	env.createRepo("web")
-	env.createWorkspace("ws-a", "feat/a", []string{"api"})
-	env.createWorkspace("ws-b", "feat/b", []string{"web"})
-
-	err := env.svc.Rename("ws-a", "ws-b")
-	if err == nil {
-		t.Error("expected error for taken name")
-	}
-}
-
-func TestRenamePreservesCreatedAt(t *testing.T) {
-	env := setupTestEnv(t)
-	env.createRepo("api")
-	env.createWorkspace("preserve-ws", "feat/preserve", []string{"api"})
-
-	ws, _ := env.svc.State.GetWorkspace("preserve-ws")
-	originalCreatedAt := ws.CreatedAt
-
-	env.svc.Rename("preserve-ws", "renamed-ws")
-
-	ws, _ = env.svc.State.GetWorkspace("renamed-ws")
-	if ws.CreatedAt != originalCreatedAt {
-		t.Errorf("created_at changed: %q -> %q", originalCreatedAt, ws.CreatedAt)
-	}
-}
-
-// TestReplaceSequence covers the Delete→Create flow that `gw create --replace`
-// performs. The key invariants: after delete, the old workspace is gone from
-// state and disk, and the freed branch can be reused immediately by the new
-// workspace without hitting the duplicate-branch guard.
-func TestReplaceSequence(t *testing.T) {
-	env := setupTestEnv(t)
-	env.createRepo("api")
-
-	// Create old workspace on a branch.
-	if err := env.createWorkspace("old-ws", "feat/shared", []string{"api"}); err != nil {
-		t.Fatalf("create old: %v", err)
-	}
-
-	// Sanity: creating another workspace on the same branch is normally rejected.
-	if err := env.createWorkspace("other", "feat/shared", []string{"api"}); err == nil {
-		t.Fatal("expected duplicate-branch rejection before delete")
-	}
-
-	// Delete old (simulates --replace first half).
-	if err := env.svc.Delete("old-ws"); err != nil {
-		t.Fatalf("delete old: %v", err)
-	}
-
-	// Old workspace is gone from state.
-	if ws, _ := env.svc.State.GetWorkspace("old-ws"); ws != nil {
-		t.Error("old-ws still in state after delete")
-	}
-	// Old workspace directory is gone from disk.
-	if _, err := os.Stat(filepath.Join(env.wsDir, "old-ws")); !os.IsNotExist(err) {
-		t.Error("old-ws directory still on disk after delete")
-	}
-
-	// Create new workspace on the SAME branch — should now succeed because
-	// the old worktree releasing the branch is the whole point of --replace.
-	if err := env.createWorkspace("new-ws", "feat/shared", []string{"api"}); err != nil {
-		t.Fatalf("create new (branch reuse after delete): %v", err)
-	}
-
-	ws, _ := env.svc.State.GetWorkspace("new-ws")
-	if ws == nil {
-		t.Fatal("new-ws not in state")
-	}
-	if len(ws.Repos) != 1 || ws.Repos[0].Branch != "feat/shared" {
-		t.Errorf("new-ws not on expected branch: %+v", ws.Repos)
-	}
-	if _, err := os.Stat(filepath.Join(env.wsDir, "new-ws", "api")); err != nil {
-		t.Errorf("new-ws worktree missing: %v", err)
-	}
-}
-
-// ---------------------------------------------------------------------------
-// Sync tests
-// ---------------------------------------------------------------------------
-
-func TestSyncUpToDate(t *testing.T) {
-	env := setupTestEnv(t)
-	env.createRepoWithRemote("api")
-	env.createWorkspace("sync-ws", "feat/sync", []string{"api"})
-
-	// No upstream changes — should be up to date
-	err := env.svc.Sync("sync-ws")
-	if err != nil {
-		t.Fatalf("sync: %v", err)
-	}
-}
-
-func TestSyncNotFound(t *testing.T) {
-	env := setupTestEnv(t)
-	_ = env
-
-	err := env.svc.Sync("nonexistent")
-	if err == nil {
-		t.Error("expected error")
-	}
-}
-
-// ---------------------------------------------------------------------------
-// AddRepos tests
-// ---------------------------------------------------------------------------
-
-func TestAddReposSuccess(t *testing.T) {
-	env := setupTestEnv(t)
-	env.createRepo("api")
-	env.createRepo("web")
-
-	env.createWorkspace("add-ws", "feat/add", []string{"api"})
-
-	err := env.svc.AddRepos("add-ws", []string{"web"}, env.repoMap)
-	if err != nil {
-		t.Fatalf("add: %v", err)
-	}
-
-	ws, _ := env.svc.State.GetWorkspace("add-ws")
-	if len(ws.Repos) != 2 {
-		t.Errorf("expected 2 repos, got %d", len(ws.Repos))
-	}
-
-	// Worktree dir exists
-	if _, err := os.Stat(filepath.Join(env.wsDir, "add-ws", "web")); os.IsNotExist(err) {
-		t.Error("web worktree dir missing")
-	}
-}
-
-func TestAddReposAlreadyPresent(t *testing.T) {
-	env := setupTestEnv(t)
-	env.createRepo("api")
-	env.createWorkspace("dup-ws", "feat/dup", []string{"api"})
-
-	// Adding same repo again should be a no-op without wait progress.
-	output := captureStderr(t, func() {
-		if err := env.svc.AddRepos("dup-ws", []string{"api"}, env.repoMap); err != nil {
-			t.Fatalf("add duplicate: %v", err)
-		}
-	})
-	if strings.Contains(output, "Adding") {
-		t.Errorf("no-op addition should not show wait progress, got: %q", output)
-	}
-
-	ws, _ := env.svc.State.GetWorkspace("dup-ws")
-	if len(ws.Repos) != 1 {
-		t.Errorf("expected 1 repo (no dup), got %d", len(ws.Repos))
-	}
-}
-
-func TestAddReposNotFound(t *testing.T) {
-	env := setupTestEnv(t)
-	_ = env
-
-	err := env.svc.AddRepos("nonexistent", []string{"api"}, env.repoMap)
-	if err == nil {
-		t.Error("expected error")
-	}
-}
-
-func TestAddReposRollsBackEarlierAdditions(t *testing.T) {
-	env := setupTestEnv(t)
-	env.createRepo("api")
-	env.createRepo("web")
-	worker := env.createRepo("worker")
-	if err := env.createWorkspace("add-rollback", "feat/add-rollback", []string{"api"}); err != nil {
-		t.Fatalf("create: %v", err)
-	}
-
-	conflictPath := filepath.Join(env.dir, "worker-conflict")
-	env.run(worker, "git", "worktree", "add", "-q", "-b", "feat/add-rollback", conflictPath)
-
-	err := env.svc.AddRepos("add-rollback", []string{"web", "worker"}, env.repoMap)
-	if err == nil || !strings.Contains(err.Error(), "worker") {
-		t.Fatalf("expected worker conflict, got %v", err)
-	}
-
-	ws, _ := env.svc.State.GetWorkspace("add-rollback")
-	if ws.FindRepo("web") != nil {
-		t.Fatal("rolled-back web repo should not be saved")
-	}
-	if _, err := os.Stat(filepath.Join(ws.Path, "web")); !os.IsNotExist(err) {
-		t.Error("rolled-back web worktree should be removed")
-	}
-	if gitops.BranchExists(env.repoMap["web"], "feat/add-rollback") {
-		t.Error("branch created for rolled-back web repo should be removed")
-	}
-}
-
-func TestAddReposRollbackPreservesPreexistingBranch(t *testing.T) {
-	env := setupTestEnv(t)
-	env.createRepo("api")
-	web := env.createRepo("web")
-	worker := env.createRepo("worker")
-	if err := env.createWorkspace("add-preserve", "feat/add-preserve", []string{"api"}); err != nil {
-		t.Fatalf("create: %v", err)
-	}
-	env.run(web, "git", "branch", "feat/add-preserve")
-	conflictPath := filepath.Join(env.dir, "worker-preserve-conflict")
-	env.run(worker, "git", "worktree", "add", "-q", "-b", "feat/add-preserve", conflictPath)
-
-	err := env.svc.AddRepos("add-preserve", []string{"web", "worker"}, env.repoMap)
-	if err == nil {
-		t.Fatal("expected worker conflict")
-	}
-	if _, err := os.Stat(filepath.Join(env.wsDir, "add-preserve", "web")); !os.IsNotExist(err) {
-		t.Error("rolled-back web worktree should be removed")
-	}
-	if !gitops.BranchExists(web, "feat/add-preserve") {
-		t.Error("rollback should preserve a branch it did not create")
-	}
-}
-
-func TestAddReposCleansBranchWhenWorktreeAddFails(t *testing.T) {
-	env := setupTestEnv(t)
-	env.createRepo("api")
-	web := env.createRepo("web")
-	if err := env.createWorkspace("add-provision-fail", "feat/add-provision-fail", []string{"api"}); err != nil {
-		t.Fatalf("create: %v", err)
-	}
-
-	blockedPath := filepath.Join(env.wsDir, "add-provision-fail", "web")
-	if err := os.MkdirAll(blockedPath, 0o755); err != nil {
-		t.Fatalf("create blocked path: %v", err)
-	}
-	if err := os.WriteFile(filepath.Join(blockedPath, "keep.txt"), []byte("occupied"), 0o644); err != nil {
-		t.Fatalf("write blocked path: %v", err)
-	}
-
-	if err := env.svc.AddRepos("add-provision-fail", []string{"web"}, env.repoMap); err == nil {
-		t.Fatal("expected worktree add failure")
-	}
-	if gitops.BranchExists(web, "feat/add-provision-fail") {
-		t.Error("branch created before failed worktree add should be cleaned up")
-	}
-	if _, err := os.Stat(filepath.Join(blockedPath, "keep.txt")); err != nil {
-		t.Fatal("pre-existing blocked path should remain untouched")
-	}
-}
-
-// ---------------------------------------------------------------------------
-// RemoveRepos tests
-// ---------------------------------------------------------------------------
-
-func TestRemoveReposSuccess(t *testing.T) {
-	env := setupTestEnv(t)
-	env.createRepo("api")
-	env.createRepo("web")
-	env.createWorkspace("rm-ws", "feat/rm", []string{"api", "web"})
-
-	err := env.svc.RemoveRepos("rm-ws", []string{"web"})
-	if err != nil {
-		t.Fatalf("remove: %v", err)
-	}
-
-	ws, _ := env.svc.State.GetWorkspace("rm-ws")
-	if len(ws.Repos) != 1 {
-		t.Errorf("expected 1 repo, got %d", len(ws.Repos))
-	}
-	if ws.Repos[0].RepoName != "api" {
-		t.Errorf("remaining repo should be api, got %s", ws.Repos[0].RepoName)
-	}
-
-	// Worktree dir removed
-	if _, err := os.Stat(filepath.Join(env.wsDir, "rm-ws", "web")); !os.IsNotExist(err) {
-		t.Error("web worktree dir should be removed")
-	}
-}
-
-func TestRemoveReposMultiple(t *testing.T) {
-	env := setupTestEnv(t)
-	env.createRepo("api")
-	env.createRepo("web")
-	env.createRepo("worker")
-	env.createWorkspace("rm-multi", "feat/rm-multi", []string{"api", "web", "worker"})
-
-	err := env.svc.RemoveRepos("rm-multi", []string{"web", "worker"})
-	if err != nil {
-		t.Fatalf("remove: %v", err)
-	}
-
-	ws, _ := env.svc.State.GetWorkspace("rm-multi")
-	if len(ws.Repos) != 1 {
-		t.Errorf("expected 1 repo remaining, got %d", len(ws.Repos))
-	}
-	if ws.Repos[0].RepoName != "api" {
-		t.Errorf("remaining should be api, got %s", ws.Repos[0].RepoName)
-	}
-
-	// Both worktree dirs should be gone
-	for _, name := range []string{"web", "worker"} {
-		wt := filepath.Join(env.wsDir, "rm-multi", name)
-		if _, err := os.Stat(wt); !os.IsNotExist(err) {
-			t.Errorf("worktree %s should be removed", name)
-		}
-	}
-}
-
-func TestRemoveReposNonexistent(t *testing.T) {
-	env := setupTestEnv(t)
-	env.createRepo("api")
-	env.createWorkspace("rm-ne", "feat/rm-ne", []string{"api"})
-
-	// Removing a repo not in workspace should be a silent no-op.
-	output := captureStderr(t, func() {
-		if err := env.svc.RemoveRepos("rm-ne", []string{"nonexistent"}); err != nil {
-			t.Fatalf("remove nonexistent: %v", err)
-		}
-	})
-	if output != "" {
-		t.Errorf("no-op removal should be silent, got: %q", output)
-	}
-}
-
-func TestRemoveReposRejectsDirtyWorktreeByDefault(t *testing.T) {
-	env := setupTestEnv(t)
-	env.createRepo("api")
-	env.createRepo("web")
-	if err := env.createWorkspace("dirty-remove", "feat/dirty-remove", []string{"api", "web"}); err != nil {
-		t.Fatalf("create: %v", err)
-	}
-
-	ws, _ := env.svc.State.GetWorkspace("dirty-remove")
-	web := ws.FindRepo("web")
-	if err := os.WriteFile(filepath.Join(web.WorktreePath, "dirty.txt"), []byte("keep me"), 0o644); err != nil {
-		t.Fatalf("write dirty file: %v", err)
-	}
-
-	err := env.svc.RemoveRepos("dirty-remove", []string{"web"})
-	if err == nil || !strings.Contains(err.Error(), "web") || !strings.Contains(err.Error(), "dirty") {
-		t.Fatalf("expected repo-named dirty error, got %v", err)
-	}
-	if saved, _ := env.svc.State.GetWorkspace("dirty-remove"); saved == nil || saved.FindRepo("web") == nil {
-		t.Fatal("dirty repo should remain in state")
-	}
-}
-
-func TestRemoveReposRetainsOnlyFailedEntries(t *testing.T) {
-	env := setupTestEnv(t)
-	env.createRepo("api")
-	env.createRepo("web")
-	env.createRepo("worker")
-	if err := env.createWorkspace("partial-remove", "feat/partial-remove", []string{"api", "web", "worker"}); err != nil {
-		t.Fatalf("create: %v", err)
-	}
-
-	ws, _ := env.svc.State.GetWorkspace("partial-remove")
-	apiPath := ws.FindRepo("api").WorktreePath
-	webPath := ws.FindRepo("web").WorktreePath
-	env.svc.RemoveWorktree = func(repo, path string, force bool) error {
-		if path == webPath {
-			return fmt.Errorf("simulated git failure")
-		}
-		return gitops.WorktreeRemove(repo, path, force)
-	}
-
-	err := env.svc.RemoveRepos("partial-remove", []string{"api", "web"})
-	if err == nil || !strings.Contains(err.Error(), "web") {
-		t.Fatalf("expected web removal error, got %v", err)
-	}
-	if _, err := os.Stat(apiPath); !os.IsNotExist(err) {
-		t.Error("successfully removed api worktree should be gone")
-	}
-	if _, err := os.Stat(webPath); err != nil {
-		t.Fatalf("failed web worktree should remain: %v", err)
-	}
-
-	saved, _ := env.svc.State.GetWorkspace("partial-remove")
-	if saved == nil || saved.FindRepo("api") != nil || saved.FindRepo("web") == nil || saved.FindRepo("worker") == nil {
-		t.Fatalf("state should retain failed and untouched repos: %+v", saved)
-	}
-}
-
-// ---------------------------------------------------------------------------
-// Status tests
-// ---------------------------------------------------------------------------
-
-func TestFormatSourceLine(t *testing.T) {
-	tests := []struct {
-		name string
-		src  *models.WorkspaceSource
-		want string
-	}{
-		{"nil", nil, ""},
-		{
-			"full",
-			&models.WorkspaceSource{Provider: "github", Ref: "1172", Title: "Surface data source status", URL: "https://github.com/o/r/pull/1172"},
-			"Source: github 1172 — Surface data source status  (https://github.com/o/r/pull/1172)",
-		},
-		{
-			"url only",
-			&models.WorkspaceSource{Provider: "slack", URL: "https://x.slack.com/archives/C/p1"},
-			"Source: slack  (https://x.slack.com/archives/C/p1)",
-		},
-		{
-			"no provider",
-			&models.WorkspaceSource{Title: "untitled"},
-			"Source: source — untitled",
-		},
-	}
-	for _, tt := range tests {
-		t.Run(tt.name, func(t *testing.T) {
-			if got := formatSourceLine(tt.src); got != tt.want {
-				t.Errorf("formatSourceLine() = %q, want %q", got, tt.want)
-			}
-		})
-	}
-}
-
-func TestStatusSuccess(t *testing.T) {
-	env := setupTestEnv(t)
-	env.createRepo("api")
-	env.createWorkspace("status-ws", "feat/status", []string{"api"})
-
-	err := env.svc.Status("status-ws", StatusOptions{})
-	if err != nil {
-		t.Fatalf("status: %v", err)
-	}
-}
-
-func TestCollectRepoStatusReportsCurrentBranch(t *testing.T) {
-	env := setupTestEnv(t)
-	env.createRepo("api")
-	env.createWorkspace("status-ws", "feat/status", []string{"api"})
-
-	ws, err := env.svc.State.GetWorkspace("status-ws")
-	if err != nil {
-		t.Fatalf("get workspace: %v", err)
-	}
-	repo := ws.Repos[0]
-	env.run(repo.WorktreePath, "git", "switch", "-q", "-c", "feat/actual")
-
-	result := collectRepoStatus(repo)
-	if result.Branch != "feat/actual" {
-		t.Errorf("branch: got %q, want %q", result.Branch, "feat/actual")
-	}
-}
-
-func TestCollectRepoStatusReportsDetachedHead(t *testing.T) {
-	env := setupTestEnv(t)
-	env.createRepo("api")
-	env.createWorkspace("status-ws", "feat/status", []string{"api"})
-
-	ws, err := env.svc.State.GetWorkspace("status-ws")
-	if err != nil {
-		t.Fatalf("get workspace: %v", err)
-	}
-	repo := ws.Repos[0]
-	env.run(repo.WorktreePath, "git", "switch", "-q", "--detach")
-
-	result := collectRepoStatus(repo)
-	if result.Branch != "(detached)" {
-		t.Errorf("branch: got %q, want %q", result.Branch, "(detached)")
-	}
-}
-
-func TestStatusJSON(t *testing.T) {
-	env := setupTestEnv(t)
-	env.createRepo("api")
-	env.createWorkspace("json-ws", "feat/json", []string{"api"})
-
-	err := env.svc.Status("json-ws", StatusOptions{JSON: true})
-	if err != nil {
-		t.Fatalf("status json: %v", err)
-	}
-}
-
-func TestStatusVerbose(t *testing.T) {
-	env := setupTestEnv(t)
-	env.createRepo("api")
-	env.createWorkspace("verbose-ws", "feat/verbose", []string{"api"})
-
-	// Should not error — verbose shows raw git status for dirty repos
-	err := env.svc.Status("verbose-ws", StatusOptions{Verbose: true})
-	if err != nil {
-		t.Fatalf("status verbose: %v", err)
-	}
-}
-
-func TestStatusNotFound(t *testing.T) {
-	env := setupTestEnv(t)
-	_ = env
-
-	err := env.svc.Status("nonexistent", StatusOptions{})
-	if err == nil {
-		t.Error("expected error")
-	}
-}
-
-// ---------------------------------------------------------------------------
-// Doctor tests
-// ---------------------------------------------------------------------------
-
-func TestDoctorHealthy(t *testing.T) {
-	env := setupTestEnv(t)
-	env.createRepo("api")
-	env.createWorkspace("healthy-ws", "feat/healthy", []string{"api"})
-
-	issues, _, err := env.svc.Doctor(false)
-	if err != nil {
-		t.Fatalf("doctor: %v", err)
-	}
-	if len(issues) != 0 {
-		t.Errorf("expected 0 issues, got %d", len(issues))
-	}
-}
-
-func TestDoctorDetectsMissingWorktree(t *testing.T) {
-	env := setupTestEnv(t)
-	env.createRepo("api")
-	env.createWorkspace("stale-ws", "feat/stale", []string{"api"})
-
-	// Delete worktree dir manually
-	os.RemoveAll(filepath.Join(env.wsDir, "stale-ws", "api"))
-
-	issues, _, err := env.svc.Doctor(false)
-	if err != nil {
-		t.Fatalf("doctor: %v", err)
-	}
-	if len(issues) == 0 {
-		t.Error("expected at least 1 issue for missing worktree")
-	}
-}
-
-func TestDoctorFixRemovesStale(t *testing.T) {
-	env := setupTestEnv(t)
-	env.createRepo("api")
-	env.createWorkspace("fix-ws", "feat/fix", []string{"api"})
-
-	// Delete worktree dir
-	os.RemoveAll(filepath.Join(env.wsDir, "fix-ws", "api"))
-
-	_, fixed, err := env.svc.Doctor(true)
-	if err != nil {
-		t.Fatalf("doctor fix: %v", err)
-	}
-	if fixed == 0 {
-		t.Error("expected at least 1 fix")
-	}
-
-	// After fix, should be clean (or at least fewer issues)
-	issues, _, _ := env.svc.Doctor(false)
-	if len(issues) > 0 {
-		// Workspace with no repos might still be an issue, that's ok
-		t.Logf("remaining issues after fix: %d", len(issues))
-	}
-}
-
-func TestDoctorDetectsMissingWorkspaceDir(t *testing.T) {
-	env := setupTestEnv(t)
-	env.createRepo("api")
-	env.createWorkspace("ghost-ws", "feat/ghost", []string{"api"})
-
-	// Delete entire workspace directory
-	os.RemoveAll(filepath.Join(env.wsDir, "ghost-ws"))
-
-	issues, _, err := env.svc.Doctor(false)
-	if err != nil {
-		t.Fatalf("doctor: %v", err)
-	}
-
-	found := false
-	for _, issue := range issues {
-		if strings.Contains(issue.Issue, "workspace directory missing") {
-			found = true
-		}
-	}
-	if !found {
-		t.Error("expected 'workspace directory missing' issue")
-	}
-}
-
-// ---------------------------------------------------------------------------
-// Parallel behavior tests
-// ---------------------------------------------------------------------------
-
-func TestCreateMultiRepoAllProcessed(t *testing.T) {
-	env := setupTestEnv(t)
-	env.createRepo("api")
-	env.createRepo("web")
-	env.createRepo("worker")
-
-	err := env.createWorkspace("multi-ws", "feat/multi", []string{"api", "web", "worker"})
-	if err != nil {
-		t.Fatalf("create: %v", err)
-	}
-
-	ws, _ := env.svc.State.GetWorkspace("multi-ws")
-	if len(ws.Repos) != 3 {
-		t.Errorf("expected 3 repos, got %d", len(ws.Repos))
-	}
-
-	// All worktrees exist
-	for _, name := range []string{"api", "web", "worker"} {
-		wt := filepath.Join(env.wsDir, "multi-ws", name)
-		if _, err := os.Stat(wt); os.IsNotExist(err) {
-			t.Errorf("worktree %s missing", name)
-		}
-	}
-}
-
-func TestStatusMultiRepoAllReported(t *testing.T) {
-	env := setupTestEnv(t)
-	env.createRepo("api")
-	env.createRepo("web")
-	env.createWorkspace("multi-status", "feat/ms", []string{"api", "web"})
-
-	// Should not error even with multiple repos
-	err := env.svc.Status("multi-status", StatusOptions{})
-	if err != nil {
-		t.Fatalf("status: %v", err)
-	}
-}
-
-func TestDeleteMultiRepoAllCleaned(t *testing.T) {
-	env := setupTestEnv(t)
-	env.createRepo("api")
-	env.createRepo("web")
-	env.createWorkspace("multi-del", "feat/md", []string{"api", "web"})
-
-	err := env.svc.Delete("multi-del")
-	if err != nil {
-		t.Fatalf("delete: %v", err)
-	}
-
-	// Both worktree dirs cleaned
-	for _, name := range []string{"api", "web"} {
-		wt := filepath.Join(env.wsDir, "multi-del", name)
-		if _, err := os.Stat(wt); !os.IsNotExist(err) {
-			t.Errorf("worktree %s should be removed", name)
-		}
-	}
-
-	// State cleared
-	ws, _ := env.svc.State.GetWorkspace("multi-del")
-	if ws != nil {
-		t.Error("workspace should be removed from state")
-	}
-}
-
-func TestSyncMultiRepo(t *testing.T) {
-	env := setupTestEnv(t)
-	env.createRepoWithRemote("api")
-	env.createRepoWithRemote("web")
-	env.createWorkspace("sync-multi", "feat/sm", []string{"api", "web"})
-
-	err := env.svc.Sync("sync-multi")
-	if err != nil {
-		t.Fatalf("sync: %v", err)
-	}
-}
-
-// ---------------------------------------------------------------------------
-// Sync with upstream changes
-// ---------------------------------------------------------------------------
-
-func TestSyncRebases(t *testing.T) {
-	env := setupTestEnv(t)
-	repo := env.createRepoWithRemote("api")
-	env.createWorkspace("rebase-ws", "feat/rebase", []string{"api"})
-	wt := filepath.Join(env.wsDir, "rebase-ws", "api")
-
-	// Add a commit upstream on main and push to origin
-	os.WriteFile(filepath.Join(repo, "upstream.txt"), []byte("new"), 0o644)
-	env.run(repo, "git", "add", ".")
-	env.run(repo, "git", "commit", "-q", "-m", "upstream change")
-	env.run(repo, "git", "push", "-q", "origin", "HEAD")
-
-	// Sync should rebase
-	err := env.svc.Sync("rebase-ws")
-	if err != nil {
-		t.Fatalf("sync: %v", err)
-	}
-
-	// The upstream file should now be in the worktree
-	if _, err := os.Stat(filepath.Join(wt, "upstream.txt")); os.IsNotExist(err) {
-		t.Error("upstream change should be rebased into worktree")
-	}
-}
-
-// ---------------------------------------------------------------------------
-// AllWorkspacesSummary tests
-// ---------------------------------------------------------------------------
-
-func TestAllWorkspacesSummaryEmpty(t *testing.T) {
-	env := setupTestEnv(t)
-	_ = env
-
-	results, err := env.svc.AllWorkspacesSummary()
-	if err != nil {
-		t.Fatalf("summary: %v", err)
-	}
-	if len(results) != 0 {
-		t.Errorf("expected empty, got %d", len(results))
-	}
-}
-
-func TestAllWorkspacesSummaryMultiple(t *testing.T) {
-	env := setupTestEnv(t)
-	env.createRepo("api")
-	env.createRepo("web")
-	env.createWorkspace("ws-a", "feat/a", []string{"api"})
-	env.createWorkspace("ws-b", "feat/b", []string{"web"})
-
-	results, err := env.svc.AllWorkspacesSummary()
-	if err != nil {
-		t.Fatalf("summary: %v", err)
-	}
-	if len(results) != 2 {
-		t.Errorf("expected 2, got %d", len(results))
-	}
-
-	// Each result should have name, branch, repos count, status, path
-	for _, r := range results {
-		if r.Name == "" || r.Branch == "" || r.Path == "" {
-			t.Errorf("empty field in result: %+v", r)
-		}
-	}
-}
-
-func TestSyncSkipsDirty(t *testing.T) {
-	env := setupTestEnv(t)
-	env.createRepoWithRemote("api")
-	env.createWorkspace("dirty-ws", "feat/dirty", []string{"api"})
-
-	wt := filepath.Join(env.wsDir, "dirty-ws", "api")
-	os.WriteFile(filepath.Join(wt, "dirt.txt"), []byte("uncommitted"), 0o644)
-
-	err := env.svc.Sync("dirty-ws")
-	if err != nil {
-		t.Fatalf("sync: %v", err)
-	}
-	// No error = correctly skipped
-}
-
-// ---------------------------------------------------------------------------
-// Teardown hook tests
-// ---------------------------------------------------------------------------
-
-func TestDeleteRunsTeardownHook(t *testing.T) {
-	env := setupTestEnv(t)
-	repo := env.createRepo("api")
-
-	// Write .grove.toml with teardown hook
-	toml := `teardown = "touch /tmp/grove-teardown-test-marker"`
-	os.WriteFile(filepath.Join(repo, ".grove.toml"), []byte(toml), 0o644)
-	env.run(repo, "git", "add", ".")
-	env.run(repo, "git", "commit", "-q", "-m", "add teardown")
-
-	// Override RunCmdSilent to capture calls
-	var teardownCalls []string
-	origRunCmdSilent := env.svc.RunCmdSilent
-	env.svc.RunCmdSilent = func(dir, cmd string) error {
-		teardownCalls = append(teardownCalls, cmd)
-		return nil
-	}
-	defer func() { env.svc.RunCmdSilent = origRunCmdSilent }()
-
-	env.createWorkspace("td-ws", "feat/td", []string{"api"})
-	env.svc.Delete("td-ws")
-
-	found := false
-	for _, c := range teardownCalls {
-		if strings.Contains(c, "teardown-test-marker") {
-			found = true
-		}
-	}
-	if !found {
-		t.Errorf("teardown hook not called; calls: %v", teardownCalls)
-	}
-}
-
-func TestRemoveReposRunsTeardownHook(t *testing.T) {
-	env := setupTestEnv(t)
-	repo := env.createRepo("api")
-	env.createRepo("web")
-
-	toml := `teardown = "echo tearing-down"`
-	os.WriteFile(filepath.Join(repo, ".grove.toml"), []byte(toml), 0o644)
-	env.run(repo, "git", "add", ".")
-	env.run(repo, "git", "commit", "-q", "-m", "add teardown")
-
-	var teardownCalls []string
-	origRunCmdSilent := env.svc.RunCmdSilent
-	env.svc.RunCmdSilent = func(dir, cmd string) error {
-		teardownCalls = append(teardownCalls, cmd)
-		return nil
-	}
-	defer func() { env.svc.RunCmdSilent = origRunCmdSilent }()
-
-	env.createWorkspace("td-rm-ws", "feat/td-rm", []string{"api", "web"})
-	env.svc.RemoveRepos("td-rm-ws", []string{"api"})
-
-	found := false
-	for _, c := range teardownCalls {
-		if strings.Contains(c, "tearing-down") {
-			found = true
-		}
-	}
-	if !found {
-		t.Errorf("teardown hook not called during remove; calls: %v", teardownCalls)
-	}
-}
-
-// ---------------------------------------------------------------------------
-// Pre/post sync hook tests
-// ---------------------------------------------------------------------------
-
-func TestSyncRunsPreAndPostHooks(t *testing.T) {
-	env := setupTestEnv(t)
-	repo := env.createRepoWithRemote("api")
-
-	toml := `pre_sync = "echo pre"
-post_sync = "echo post"`
-	os.WriteFile(filepath.Join(repo, ".grove.toml"), []byte(toml), 0o644)
-	env.run(repo, "git", "add", ".")
-	env.run(repo, "git", "commit", "-q", "-m", "add sync hooks")
-	env.run(repo, "git", "push", "-q", "origin", "HEAD")
-
-	env.createWorkspace("sync-hook-ws", "feat/sync-hook", []string{"api"})
-
-	// Add upstream commit to trigger rebase
-	os.WriteFile(filepath.Join(repo, "new.txt"), []byte("upstream"), 0o644)
-	env.run(repo, "git", "add", ".")
-	env.run(repo, "git", "commit", "-q", "-m", "upstream")
-	env.run(repo, "git", "push", "-q", "origin", "HEAD")
-
-	var hookCalls []string
-	origRunCmdSilent := env.svc.RunCmdSilent
-	env.svc.RunCmdSilent = func(dir, cmd string) error {
-		hookCalls = append(hookCalls, cmd)
-		return nil
-	}
-	defer func() { env.svc.RunCmdSilent = origRunCmdSilent }()
-
-	env.svc.Sync("sync-hook-ws")
-
-	foundPre := false
-	foundPost := false
-	for _, c := range hookCalls {
-		if strings.Contains(c, "pre") {
-			foundPre = true
-		}
-		if strings.Contains(c, "post") {
-			foundPost = true
-		}
-	}
-	if !foundPre {
-		t.Errorf("pre_sync hook not called; calls: %v", hookCalls)
-	}
-	if !foundPost {
-		t.Errorf("post_sync hook not called; calls: %v", hookCalls)
-	}
-}
-
-// ---------------------------------------------------------------------------
-// Setup hook injection test
-// ---------------------------------------------------------------------------
-
-func TestSetupHookUsesRunCmd(t *testing.T) {
-	env := setupTestEnv(t)
-	repo := env.createRepo("api")
-
-	toml := `setup = "echo injected-setup"`
-	os.WriteFile(filepath.Join(repo, ".grove.toml"), []byte(toml), 0o644)
-	env.run(repo, "git", "add", ".")
-	env.run(repo, "git", "commit", "-q", "-m", "add setup")
-
-	var setupCalls []string
-	origRunCmd := env.svc.RunCmd
-	env.svc.RunCmd = func(dir, cmd string) error {
-		setupCalls = append(setupCalls, cmd)
-		return nil
-	}
-	defer func() { env.svc.RunCmd = origRunCmd }()
-
-	env.createWorkspace("inject-ws", "feat/inject", []string{"api"})
-
-	found := false
-	for _, c := range setupCalls {
-		if strings.Contains(c, "injected-setup") {
-			found = true
-		}
-	}
-	if !found {
-		t.Errorf("setup hook should use RunCmd; calls: %v", setupCalls)
-	}
-}
-
-// ---------------------------------------------------------------------------
-// Delete partial failure preserves state
-// ---------------------------------------------------------------------------
-
-func TestDeletePartialFailurePreservesState(t *testing.T) {
-	env := setupTestEnv(t)
-	env.createRepo("api")
-	env.createRepo("web")
-	env.createWorkspace("partial-ws", "feat/partial", []string{"api", "web"})
-
-	// Corrupt one worktree by removing its .git reference
-	// This simulates a worktree that can't be removed via git
-	ws, _ := env.svc.State.GetWorkspace("partial-ws")
-	apiWT := ws.Repos[0].WorktreePath
-
-	// Lock the worktree dir to prevent removal (create a dir that looks like it's still there)
-	// We can't easily simulate a hard failure, but we can verify the state
-	// by checking that the dir exists after delete
-	_ = apiWT
-
-	// The existing Delete implementation removes the directory regardless,
-	// so we test that state is removed when everything succeeds
-	env.svc.Delete("partial-ws")
-
-	wsAfter, _ := env.svc.State.GetWorkspace("partial-ws")
-	if wsAfter != nil {
-		t.Error("on successful delete, workspace should be removed from state")
-	}
-}
-
-// ---------------------------------------------------------------------------
-// AddRepos edge cases
-// ---------------------------------------------------------------------------
-
-func TestAddReposBranchConflict(t *testing.T) {
-	env := setupTestEnv(t)
-	env.createRepo("api")
-	env.createRepo("web")
-
-	// Create first workspace, consuming the branch on "web"
-	env.createWorkspace("ws1", "feat/conflict", []string{"web"})
-
-	// Create second workspace with "api" only
-	env.createWorkspace("ws2", "feat/other", []string{"api"})
-
-	// Try to add "web" to ws2 with a different branch — but web already has
-	// feat/conflict. This should work because it's a different branch.
-	// But adding web with a branch that already has a worktree should fail.
-	// The branch "feat/conflict" already has a worktree, so adding it again should error.
-	err := env.svc.AddRepos("ws2", []string{"web"}, env.repoMap)
-	// This will try to create branch "feat/other" on "web" — should work (different branch)
-	if err != nil {
-		t.Fatalf("adding web with different branch should succeed: %v", err)
-	}
-}
-
-func TestAddReposRunsSetupHooks(t *testing.T) {
-	env := setupTestEnv(t)
-	env.createRepo("api")
-	repo := env.createRepo("web")
-
-	toml := `setup = "touch .added-setup"`
-	os.WriteFile(filepath.Join(repo, ".grove.toml"), []byte(toml), 0o644)
-	env.run(repo, "git", "add", ".")
-	env.run(repo, "git", "commit", "-q", "-m", "add setup")
-
-	env.createWorkspace("setup-add-ws", "feat/setup-add", []string{"api"})
-
-	var setupCalls []string
-	origRunCmd := env.svc.RunCmd
-	env.svc.RunCmd = func(dir, cmd string) error {
-		setupCalls = append(setupCalls, cmd)
-		return nil
-	}
-	defer func() { env.svc.RunCmd = origRunCmd }()
-
-	env.svc.AddRepos("setup-add-ws", []string{"web"}, env.repoMap)
-
-	found := false
-	for _, c := range setupCalls {
-		if strings.Contains(c, "added-setup") {
-			found = true
-		}
-	}
-	if !found {
-		t.Errorf("setup hook should run on newly added repos; calls: %v", setupCalls)
-	}
-}
-
-// ---------------------------------------------------------------------------
-// RemoveRepos error paths
-// ---------------------------------------------------------------------------
-
-func TestRemoveReposWorkspaceNotFound(t *testing.T) {
-	env := setupTestEnv(t)
-	_ = env
-
-	err := env.svc.RemoveRepos("nonexistent", []string{"api"})
-	if err == nil {
-		t.Error("expected error for nonexistent workspace")
-	}
-}
-
-// ---------------------------------------------------------------------------
-// Sync conflict handling
-// ---------------------------------------------------------------------------
-
-// ---------------------------------------------------------------------------
-// Progress output tests
-// ---------------------------------------------------------------------------
-
-func captureStderr(t *testing.T, fn func()) string {
-	t.Helper()
-
-	oldStderr := os.Stderr
-	r, w, err := os.Pipe()
-	if err != nil {
-		t.Fatalf("capture stderr: %v", err)
-	}
-	os.Stderr = w
-	defer func() {
-		os.Stderr = oldStderr
-		w.Close()
-		r.Close()
-	}()
-
-	fn()
-	if err := w.Close(); err != nil {
-		t.Fatalf("close stderr writer: %v", err)
-	}
-
-	var buf bytes.Buffer
-	if _, err := buf.ReadFrom(r); err != nil {
-		t.Fatalf("read stderr: %v", err)
-	}
-	return buf.String()
-}
-
-func TestAddReposShowsProgressOnStderr(t *testing.T) {
-	tests := []struct {
-		name     string
-		toAdd    []string
-		expected string
-	}{
-		{name: "singular", toAdd: []string{"web"}, expected: "Adding 1 repo to add-progress-singular. Please wait."},
-		{name: "plural", toAdd: []string{"web", "worker"}, expected: "Adding 2 repos to add-progress-plural. Please wait."},
-	}
-
-	for _, tt := range tests {
-		t.Run(tt.name, func(t *testing.T) {
-			env := setupTestEnv(t)
-			env.createRepo("api")
-			env.createRepo("web")
-			env.createRepo("worker")
-
-			wsName := "add-progress-" + tt.name
-			if err := env.createWorkspace(wsName, "feat/"+wsName, []string{"api"}); err != nil {
-				t.Fatalf("create: %v", err)
-			}
-
-			output := captureStderr(t, func() {
-				if err := env.svc.AddRepos(wsName, tt.toAdd, env.repoMap); err != nil {
-					t.Fatalf("add repos: %v", err)
-				}
-			})
-			if !strings.Contains(output, tt.expected) {
-				t.Errorf("expected %q in stderr, got: %q", tt.expected, output)
-			}
-		})
-	}
-}
-
-func TestRemoveReposShowsProgressOnStderr(t *testing.T) {
-	tests := []struct {
-		name     string
-		toRemove []string
-		expected string
-	}{
-		{name: "singular", toRemove: []string{"web"}, expected: "Removing 1 repo from remove-progress-singular. Please wait."},
-		{name: "plural", toRemove: []string{"web", "worker"}, expected: "Removing 2 repos from remove-progress-plural. Please wait."},
-	}
-
-	for _, tt := range tests {
-		t.Run(tt.name, func(t *testing.T) {
-			env := setupTestEnv(t)
-			env.createRepo("api")
-			env.createRepo("web")
-			env.createRepo("worker")
-
-			wsName := "remove-progress-" + tt.name
-			if err := env.createWorkspace(wsName, "feat/"+wsName, []string{"api", "web", "worker"}); err != nil {
-				t.Fatalf("create: %v", err)
-			}
-
-			output := captureStderr(t, func() {
-				if err := env.svc.RemoveRepos(wsName, tt.toRemove); err != nil {
-					t.Fatalf("remove repos: %v", err)
-				}
-			})
-			if !strings.Contains(output, tt.expected) {
-				t.Errorf("expected %q in stderr, got: %q", tt.expected, output)
-			}
-		})
-	}
-}
-
-func TestCreateShowsProgressOnStderr(t *testing.T) {
-	env := setupTestEnv(t)
-	env.createRepo("api")
-	env.createRepo("web")
-
-	// Capture stderr
-	oldStderr := os.Stderr
-	r, w, _ := os.Pipe()
-	os.Stderr = w
-
-	env.createWorkspace("progress-ws", "feat/progress", []string{"api", "web"})
-
-	w.Close()
-	var buf bytes.Buffer
-	buf.ReadFrom(r)
-	os.Stderr = oldStderr
-
-	output := buf.String()
-
-	// Should show fetch progress then per-repo progress
-	if !strings.Contains(output, "fetching 2 repos") {
-		t.Errorf("expected 'fetching 2 repos' in output, got: %q", output)
-	}
-	if !strings.Contains(output, "[1/2]") {
-		t.Errorf("expected [1/2] progress, got: %q", output)
-	}
-	if !strings.Contains(output, "[2/2]") {
-		t.Errorf("expected [2/2] progress, got: %q", output)
-	}
-}
-
-func TestSyncConflictAbortsRebase(t *testing.T) {
-	env := setupTestEnv(t)
-	repo := env.createRepoWithRemote("api")
-	env.createWorkspace("conflict-ws", "feat/conflict", []string{"api"})
-
-	wt := filepath.Join(env.wsDir, "conflict-ws", "api")
-
-	// Make a conflicting change in the worktree
-	os.WriteFile(filepath.Join(wt, "README.md"), []byte("worktree version"), 0o644)
-	env.run(wt, "git", "add", ".")
-	env.run(wt, "git", "commit", "-q", "-m", "worktree change")
-
-	// Make a conflicting change upstream on the same file
-	os.WriteFile(filepath.Join(repo, "README.md"), []byte("upstream version"), 0o644)
-	env.run(repo, "git", "add", ".")
-	env.run(repo, "git", "commit", "-q", "-m", "upstream conflict")
-	env.run(repo, "git", "push", "-q", "origin", "HEAD")
-
-	// Sync should handle the conflict gracefully (abort rebase, no error)
-	err := env.svc.Sync("conflict-ws")
-	if err != nil {
-		t.Fatalf("sync should not return error on conflict: %v", err)
-	}
-
-	// Worktree should not be in a rebase state
-	rebaseMergeDir := filepath.Join(wt, ".git", "rebase-merge")
-	if _, err := os.Stat(rebaseMergeDir); err == nil {
-		// .git might be a file (worktree), check differently
-		t.Log("checking rebase state via git status")
-	}
-
-	// The worktree change should still be present (rebase was aborted)
-	data, _ := os.ReadFile(filepath.Join(wt, "README.md"))
-	if string(data) != "worktree version" {
-		t.Errorf("after abort, worktree should keep its version; got %q", string(data))
-	}
-}
-
-// ---------------------------------------------------------------------------
-// Logging e2e tests — verify the flight recorder captures key operations
-// ---------------------------------------------------------------------------
-
-// setupLogging points the logging package at a temp dir and returns a function
-// to read the log contents. Always call in tests that check log output.
-func setupLogging(t *testing.T) func() string {
-	t.Helper()
-	dir := t.TempDir()
-	logging.LogDir = dir
-	logging.Setup(false) // non-verbose: Info/Warn/Error still written
-
-	return func() string {
-		data, err := os.ReadFile(filepath.Join(dir, "grove.log"))
-		if err != nil {
-			t.Fatalf("reading log file: %v", err)
-		}
-		return string(data)
-	}
-}
-
-func TestLoggingCreateAndDelete(t *testing.T) {
-	readLog := setupLogging(t)
-	env := setupTestEnv(t)
-	env.createRepo("api")
-	env.createRepo("web")
-
-	// Create workspace
-	err := env.createWorkspace("log-ws", "feat/log", []string{"api", "web"})
-	if err != nil {
-		t.Fatalf("create: %v", err)
-	}
-
-	log := readLog()
-
-	// Should log creation start with workspace name, branch, and repos
-	if !strings.Contains(log, `creating workspace "log-ws"`) {
-		t.Errorf("log should contain creation start, got:\n%s", log)
-	}
-	if !strings.Contains(log, "feat/log") {
-		t.Errorf("log should contain branch name, got:\n%s", log)
-	}
-	if !strings.Contains(log, `workspace "log-ws" created`) {
-		t.Errorf("log should contain creation success, got:\n%s", log)
-	}
-
-	// Should log branch creation for each repo
-	if !strings.Contains(log, `creating branch "feat/log" in api`) {
-		t.Errorf("log should contain branch creation for api, got:\n%s", log)
-	}
-	if !strings.Contains(log, `creating branch "feat/log" in web`) {
-		t.Errorf("log should contain branch creation for web, got:\n%s", log)
-	}
-
-	// Delete workspace
-	err = env.svc.Delete("log-ws")
-	if err != nil {
-		t.Fatalf("delete: %v", err)
-	}
-
-	log = readLog()
-
-	if !strings.Contains(log, `deleting workspace "log-ws"`) {
-		t.Errorf("log should contain deletion start, got:\n%s", log)
-	}
-	if !strings.Contains(log, `workspace "log-ws" deleted`) {
-		t.Errorf("log should contain deletion success, got:\n%s", log)
-	}
-	// Branch deletion should be logged
-	if !strings.Contains(log, `deleted branch "feat/log"`) {
-		t.Errorf("log should contain branch deletion, got:\n%s", log)
-	}
-}
-
-func TestLoggingSync(t *testing.T) {
-	readLog := setupLogging(t)
-	env := setupTestEnv(t)
-	env.createRepoWithRemote("api")
-
-	err := env.createWorkspace("sync-log-ws", "feat/synclog", []string{"api"})
-	if err != nil {
-		t.Fatalf("create: %v", err)
-	}
-
-	err = env.svc.Sync("sync-log-ws")
-	if err != nil {
-		t.Fatalf("sync: %v", err)
-	}
-
-	log := readLog()
-	if !strings.Contains(log, `syncing workspace "sync-log-ws"`) {
-		t.Errorf("log should contain sync start, got:\n%s", log)
-	}
-}
-
-func TestLoggingAddAndRemoveRepos(t *testing.T) {
-	readLog := setupLogging(t)
-	env := setupTestEnv(t)
-	env.createRepo("api")
-	env.createRepo("web")
-
-	env.createWorkspace("addrem-ws", "feat/addrem", []string{"api"})
-
-	err := env.svc.AddRepos("addrem-ws", []string{"web"}, env.repoMap)
-	if err != nil {
-		t.Fatalf("add-repo: %v", err)
-	}
-
-	log := readLog()
-	if !strings.Contains(log, `added 1 repo(s) to workspace "addrem-ws"`) {
-		t.Errorf("log should contain add-repo, got:\n%s", log)
-	}
-
-	err = env.svc.RemoveRepos("addrem-ws", []string{"web"})
-	if err != nil {
-		t.Fatalf("remove-repo: %v", err)
-	}
-
-	log = readLog()
-	if !strings.Contains(log, `removed 1 repo(s) from workspace "addrem-ws"`) {
-		t.Errorf("log should contain remove-repo, got:\n%s", log)
-	}
-}
-
-func TestLoggingRename(t *testing.T) {
-	readLog := setupLogging(t)
-	env := setupTestEnv(t)
-	env.createRepo("api")
-
-	env.createWorkspace("old-name", "feat/rename", []string{"api"})
-
-	err := env.svc.Rename("old-name", "new-name")
-	if err != nil {
-		t.Fatalf("rename: %v", err)
-	}
-
-	log := readLog()
-	if !strings.Contains(log, `workspace "old-name" renamed to "new-name"`) {
-		t.Errorf("log should contain rename, got:\n%s", log)
-	}
-}
-
-func TestDeletePreservesUnmergedBranchWithoutForce(t *testing.T) {
-	readLog := setupLogging(t)
-	env := setupTestEnv(t)
-	env.createRepo("api")
-
-	env.createWorkspace("unmerged-ws", "feat/unmerged", []string{"api"})
-
-	wt := filepath.Join(env.wsDir, "unmerged-ws", "api")
-	os.WriteFile(filepath.Join(wt, "new.txt"), []byte("unmerged work"), 0o644)
-	env.run(wt, "git", "add", ".")
-	env.run(wt, "git", "commit", "-q", "-m", "unmerged commit")
-
-	sourceRepo := env.repoMap["api"]
-	if err := env.svc.Delete("unmerged-ws"); err != nil {
-		t.Fatalf("delete: %v", err)
-	}
-
-	if !gitops.BranchExists(sourceRepo, "feat/unmerged") {
-		t.Error("safe delete should preserve an unmerged branch")
-	}
-	if log := readLog(); !strings.Contains(log, "failed to delete branch") {
-		t.Errorf("log should explain preserved branch, got:\n%s", log)
-	}
-}
-
-func TestDeleteForceDeletesUnmergedBranch(t *testing.T) {
-	env := setupTestEnv(t)
-	env.createRepo("api")
-
-	env.createWorkspace("force-unmerged-ws", "feat/force-unmerged", []string{"api"})
-	wt := filepath.Join(env.wsDir, "force-unmerged-ws", "api")
-	os.WriteFile(filepath.Join(wt, "new.txt"), []byte("unmerged work"), 0o644)
-	env.run(wt, "git", "add", ".")
-	env.run(wt, "git", "commit", "-q", "-m", "unmerged commit")
-
-	if err := env.svc.DeleteWithOptions("force-unmerged-ws", RemoveOptions{Force: true}); err != nil {
-		t.Fatalf("force delete: %v", err)
-	}
-	if gitops.BranchExists(env.repoMap["api"], "feat/force-unmerged") {
-		t.Error("force delete should remove an unmerged branch")
-	}
-}
